+++ _posts/2025-06-20-helloworld.md
@@ -1,0 +1,6 @@
+---
+title: Hello world
+date: 2025-06-20
+---
+
+Wicked cool blog
