+++ app/index.css
@@ -3,6 +3,10 @@
 @import 'tailwindcss/theme.css' layer(theme);
 @import 'tailwindcss/utilities' layer(utilities);
 
+a {
+  color: #bf141c;
+}
+
 body {
   font-family:
     Fira Sans,
