+++ lib/api.ts
@@ -1,0 +1,41 @@
+import fs from 'fs'
+import path from 'path'
+import matter from 'gray-matter'
+
+const postsDirectory = path.join(process.cwd(), '_posts')
+
+function getPostFiles() {
+  return fs.readdirSync(postsDirectory)
+}
+
+export async function getPostById(id: string) {
+  const realId = id.replace(/\.md$/, '')
+  const fullPath = path.join(postsDirectory, `${realId}.md`)
+  const { data, content } = matter(await fs.promises.readFile(fullPath, 'utf8'))
+  const date = data.date as Date
+
+  return {
+    ...data,
+    title: data.title as string,
+    id: realId,
+    date: date.toISOString().slice(0, 10),
+    content,
+  }
+}
+
+export async function getPageMarkdown(string_: string) {
+  const { data, content } = matter(
+    fs.readFileSync(path.join('_pages', string_), 'utf8'),
+  )
+  const html = await parser.process(content)
+
+  return {
+    ...data,
+    html: html.value.toString(),
+  }
+}
+
+export async function getAllPosts() {
+  const posts = await Promise.all(getPostFiles().map(id => getPostById(id)))
+  return posts.sort((post1, post2) => (post1.date > post2.date ? -1 : 1))
+}
