+++ package.json
@@ -13,11 +13,13 @@
   },
   "dependencies": {
     "date-fns": "^4.1.0",
+    "gray-matter": "^4.0.3",
     "ideogram": "^1.50.0",
     "lucide-react": "^0.503.0",
     "next": "^15.0.4",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-markdown": "^10.1.0",
     "typescript": "^5.4.2",
     "usehooks-ts": "^3.1.0"
   },
